requirements.txt
# Core Framework
Flask==3.0.3
python-telegram-bot==21.4

# Firebase
firebase-admin==6.5.0
google-cloud-firestore==2.16.0

# Web Scraping & Bypass
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==2.2.2
cloudscraper==1.2.71
curl-cffi==0.7.1
selenium==4.23.1
webdriver-manager==4.0.2
playwright==1.46.0

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.3.0
html5lib==1.1

# JavaScript Execution
PyExecJS==1.5.1
demjson3==3.0.6

# AI & LLM
openai==1.42.0
langchain==0.2.14
langchain-openai==0.1.22

# Security & Encoding
pycryptodome==3.20.0
base58==2.1.1

# Async & Concurrency
aiohttp==3.10.5
asyncio==3.4.3
aiofiles==24.1.0

# Caching & Performance
cachetools==5.5.0
redis==5.0.8

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
tenacity==9.0.0
fake-useragent==1.5.1

# Monitoring & Logging
sentry-sdk==2.13.0
loguru==0.7.2

# Date & Time
python-dateutil==2.9.0
pytz==2024.1

# Validation
validators==0.33.0
pydantic==2.8.2

# Background Tasks
APScheduler==3.10.4

# Webhook Server
gunicorn==23.0.0
